requirements.txt
streamlit==1.32.0
sentence-transformers==2.3.1
pinecone-client==3.0.2
python-dotenv==1.0.0
transformers==4.36.2
torch==2.2.0
requests==2.31.0
